replace terminate by kill
... because factorio do not force stop after receiving SIGTERM on Linux
Other changes:
Try gracefully shutdown at first rather than kill when daemon is cancelled.
Thus, we should also change to directly kill the process when timeout happens, otherwise the timeout may be doubled.
Kill the old process when a daemon starts to avoid leaving the orphan processes there.
Suppress the error from send_signal and record to log.
Killed process is expected to stop soon, so adding a block argument to synchronize for done killing and log as error if not stop in 1s.